backend/requirements.txt
asgiref==3.8.1
colorama==0.4.6
Django==5.2.1
iniconfig==2.1.0
packaging==25.0
pluggy==1.5.0
pyserial==3.5
pytest==8.3.5
pytest-mock==3.14.0
python-dotenv==1.1.0
sqlparse==0.5.3
tzdata==2025.2
